Split Files
Split files for readability and ease of use / modification
